requirements.txt
fastapi==0.61.2
uvicorn==0.11.7
requests==2.19.1
PyYAML==3.13
Jinja2==2.10

